+++ tests/test_update_signalk_data.py
@@ -3,9 +3,12 @@ import sys
 import os
 from pathlib import Path
 from unittest.mock import patch, MagicMock
+import subprocess
 
 
-def test_run_on_dev_branch_writes_file_and_calls_git(tmp_path):
+
+
+def test_run_on_dev_branch_writes_file_and_calls_git(tmp_path, test_branch):
     # Fake requests module to avoid dependency and network
     class FakeResp:
         def raise_for_status(self):
@@ -49,7 +52,7 @@ def test_run_on_dev_branch_writes_file_and_calls_git(tmp_path):
         out = tmp_path / "signalk_latest.json"
 
         mod.run_update(
-            branch="_test_branch_tmp",
+            branch=test_branch,
             remote="origin",
             signalk_url="http://example",  # mocked
             output_path=str(out),
@@ -64,12 +67,12 @@ def test_run_on_dev_branch_writes_file_and_calls_git(tmp_path):
         assert json.loads(out.read_text()) == {"ok": True}
         # Ensure reset targeted origin/_test_branch_tmp and commit without --amend --force
         assert ["git", "fetch", "--all"] in calls
-        assert ["git", "reset", "--hard", "origin/_test_branch_tmp"] in calls
+        assert ["git", "reset", "--hard", f"origin/{test_branch}"] in calls
         assert any(cmd[:3] == ["git", "commit", "--allow-empty"] for cmd in calls)
         assert ["git", "push"] in calls
 
 
-def test_https_conversion(tmp_path):
+def test_https_conversion(tmp_path, test_branch):
     """Test that HTTP URLs are converted to HTTPS when use_https=True."""
     class FakeResp:
         def raise_for_status(self):
@@ -105,7 +108,7 @@ def test_https_conversion(tmp_path):
         out = tmp_path / "signalk_latest.json"
 
         mod.run_update(
-            branch="main",
+            branch=test_branch,
             remote="origin",
             signalk_url="http://example.com:3000/api",
             output_path=str(out),
@@ -120,7 +123,7 @@ def test_https_conversion(tmp_path):
         assert json.loads(out.read_text()) == {"data": "test"}
 
 
-def test_no_reset_mode(tmp_path):
+def test_no_reset_mode(tmp_path, test_branch):
     """Test that git reset is skipped when no_reset=True."""
     class FakeResp:
         def raise_for_status(self):
@@ -154,7 +157,7 @@ def test_no_reset_mode(tmp_path):
         out = tmp_path / "signalk_latest.json"
 
         mod.run_update(
-            branch="main",
+            branch=test_branch,
             remote="origin",
             signalk_url="http://example.com/api",
             output_path=str(out),
@@ -171,7 +174,7 @@ def test_no_reset_mode(tmp_path):
         assert len(reset_calls) == 0
 
 
-def test_amend_commit(tmp_path):
+def test_amend_commit(tmp_path, test_branch):
     """Test that --amend flag is used when amend=True."""
     class FakeResp:
         def raise_for_status(self):
@@ -205,7 +208,7 @@ def test_amend_commit(tmp_path):
         out = tmp_path / "signalk_latest.json"
 
         mod.run_update(
-            branch="main",
+            branch=test_branch,
             remote="origin",
             signalk_url="http://example.com/api",
             output_path=str(out),
@@ -223,7 +226,7 @@ def test_amend_commit(tmp_path):
         assert "--amend" in commit_calls[0]
 
 
-def test_force_push(tmp_path):
+def test_force_push(tmp_path, test_branch):
     """Test that --force flag is used when force_push=True."""
     class FakeResp:
         def raise_for_status(self):
@@ -257,7 +260,7 @@ def test_force_push(tmp_path):
         out = tmp_path / "signalk_latest.json"
 
         mod.run_update(
-            branch="main",
+            branch=test_branch,
             remote="origin",
             signalk_url="http://example.com/api",
             output_path=str(out),
@@ -275,7 +278,7 @@ def test_force_push(tmp_path):
         assert "--force" in push_calls[0]
 
 
-def test_no_push_mode(tmp_path):
+def test_no_push_mode(tmp_path, test_branch):
     """Test that git push is skipped when no_push=True."""
     class FakeResp:
         def raise_for_status(self):
@@ -309,7 +312,7 @@ def test_no_push_mode(tmp_path):
         out = tmp_path / "signalk_latest.json"
 
         mod.run_update(
-            branch="main",
+            branch=test_branch,
             remote="origin",
             signalk_url="http://example.com/api",
             output_path=str(out),
@@ -326,7 +329,7 @@ def test_no_push_mode(tmp_path):
         assert len(push_calls) == 0
 
 
-def test_requests_error_handling(tmp_path):
+def test_requests_error_handling(tmp_path, test_branch):
     """Test that requests errors are properly handled."""
     class FakeRequests:
         @staticmethod
@@ -356,7 +359,7 @@ def test_requests_error_handling(tmp_path):
         # Should raise an exception when requests fails
         try:
             mod.run_update(
-                branch="main",
+                branch=test_branch,
                 remote="origin",
                 signalk_url="http://example.com/api",
                 output_path=str(out),
@@ -371,3 +374,106 @@ def test_requests_error_handling(tmp_path):
             assert "Network error" in str(e)
             # Verify no file was written
             assert not out.exists()
+
+
+
+@patch.dict(sys.modules, {"requests": None})
+def test_integration_updates_test_branch(tmp_path, test_branch):
+
+    def run(cmd, cwd=None):
+        subprocess.run(cmd, check=True, cwd=cwd)
+
+    # Create a bare origin repo
+    origin = tmp_path / "origin.git"
+    origin.mkdir()
+    run(["git", "init", "--bare"], cwd=origin)
+    run(["git", "config", "user.email", "test@example.com"], cwd=origin)
+    run(["git", "config", "user.name", "Test"], cwd=origin)
+
+    # Seed repo to create branches on origin
+    seed = tmp_path / "seed"
+    seed.mkdir()
+    run(["git", "init"], cwd=seed)
+    (seed / "README.md").write_text("seed")
+    run(["git", "add", "README.md"], cwd=seed)
+    run(
+        [
+            "git",
+            "-c",
+            "user.email=test@example.com",
+            "-c",
+            "user.name=Test",
+            "commit",
+            "-m",
+            "init",
+        ],
+        cwd=seed,
+    )
+    run(["git", "remote", "add", "origin", str(origin)], cwd=seed)
+    run(["git", "push", "-u", "origin", "master"], cwd=seed)
+    # Create _test_branch_tmp branch and push
+    run(["git", "checkout", "-b", test_branch], cwd=seed)
+    run(["git", "push", "-u", "origin", test_branch], cwd=seed)
+
+    # Working repo (what the script will operate on)
+    work = tmp_path / "work"
+    run(["git", "clone", str(origin), str(work)])
+    
+    # Fetch all remote branches and checkout test_branch branch
+    run(["git", "fetch", "--all"], cwd=work)
+    run(["git", "checkout", "-b", test_branch, f"origin/{test_branch}"], cwd=work)
+
+    # Fake requests module to avoid network
+    class FakeResp:
+        def raise_for_status(self):
+            pass
+
+        def json(self):
+            return {"ok": True}
+
+    class FakeRequests:
+        @staticmethod
+        def get(url):
+            return FakeResp()
+
+    # Patch the requests module with our fake implementation
+    with patch.dict(sys.modules, {"requests": FakeRequests}):
+        # Import the script module from the project under test
+        test_file_dir = Path(__file__).parent
+        project_root = test_file_dir.parent
+        sys.path.insert(0, str(project_root))
+        import importlib
+
+        mod = importlib.import_module("scripts.update_signalk_data")
+
+        # Run from within the working repo
+        out = work / "signalk_latest.json"
+        
+        # Change to the working directory before running the script
+        original_cwd = os.getcwd()
+        try:
+            os.chdir(work)
+            mod.run_update(
+                branch=test_branch,
+                remote="origin",
+                signalk_url="http://example",
+                output_path=str(out),
+                use_https=False,
+                no_reset=False,
+                amend=False,
+                no_push=False,
+                force_push=False,
+            )
+        finally:
+            os.chdir(original_cwd)
+
+        assert out.exists()
+        # Ensure test_branch branch on origin has advanced (contains latest file content commit)
+        # Fetch the ref hash
+        before_after = (
+            subprocess.check_output(["git", "ls-remote", str(origin), f"refs/heads/{test_branch}"])
+            .decode()
+            .strip()
+        )
+        assert before_after, f"{test_branch} branch should exist on origin and have a commit"
+
